11-kedro-anti-fatigue/anti-fatigue/requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
bleach==6.2.0
build==1.2.2.post1
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.3
contourpy==1.3.3
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
duckduckgo_search==8.1.1
dynaconf==3.2.11
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.59.0
fqdn==1.5.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kedro==1.0.0
kedro-telemetry==0.6.3
kiwisolver==1.4.8
lark==1.2.2
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
omegaconf==2.3.0
opencv-contrib-python==4.12.0.88
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
zipp==3.23.0
